Move all device identification into a separate module
