+++ src/engine/strat_engine/backstore/mod.rs
@@ -9,15 +9,13 @@ mod blockdevmgr;
 mod cache_tier;
 mod data_tier;
 mod device;
+mod identify;
 mod metadata;
 mod range_alloc;
 mod setup;
 mod shared;
 
 pub use self::{
-    backstore::Backstore,
-    blockdev::StratBlockDev,
-    device::DevOwnership,
-    metadata::MDADataSize,
-    setup::{find_all, get_metadata},
+    backstore::Backstore, blockdev::StratBlockDev, device::DevOwnership, identify::find_all,
+    metadata::MDADataSize, setup::get_metadata,
 };
